+++ LinqFaster/Min.cs
@@ -29,7 +29,7 @@ namespace JM.LinqFaster
             if (r == null)
             {
                 r = source[0];
-                for (long i = 1; i < source.Length; i++)
+                for (int i = 1; i < source.Length; i++)
                 {
                     if (source[i] != null && comparer.Compare(source[i], r) < 0) r = source[i];
                 }
@@ -37,7 +37,7 @@ namespace JM.LinqFaster
             else
             {
                 r = source[0];
-                for (long i = 1; i < source.Length; i++)
+                for (int i = 1; i < source.Length; i++)
                 {
                     if (comparer.Compare(source[i], r) < 0) r = source[i];
                 }
@@ -72,7 +72,7 @@ namespace JM.LinqFaster
             if (r == null)
             {
                 r = selector(source[0]);
-                for (long i = 1; i < source.Length; i++)
+                for (int i = 1; i < source.Length; i++)
                 {
                     var v = selector(source[i]);
                     if (v != null && comparer.Compare(v, r) < 0) r = v;
@@ -81,7 +81,7 @@ namespace JM.LinqFaster
             else
             {
                 r = selector(source[0]);
-                for (long i = 1; i < source.Length; i++)
+                for (int i = 1; i < source.Length; i++)
                 {
                     var v = selector(source[i]);
                     if (comparer.Compare(v, r) < 0) r = v;
